src/test/java/org/identifiers/cloud/hq/ws/registry/models/validators/PrefixRegistrationRequestValidatorIdRegexPatternTest.java
package org.identifiers.cloud.hq.ws.registry.models.validators;

import org.identifiers.cloud.hq.ws.registry.api.requests.ServiceRequestRegisterPrefixPayload;
import org.junit.jupiter.api.Test;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.function.Function;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class PrefixRegistrationRequestValidatorIdRegexPatternTest {

    PrefixRegistrationRequestValidatorIdRegexPattern validator
            = new PrefixRegistrationRequestValidatorIdRegexPattern();

    Function<String, ServiceRequestRegisterPrefixPayload> fnc =
            (p) -> new ServiceRequestRegisterPrefixPayload()
                    .setIdRegexPattern(p);

    @Test
    void validateNullPattern() {
        assertThrows(PrefixRegistrationRequestValidatorException.class, () ->
                validator.validate(fnc.apply(null))
        );
    }

    @Test
    void validateBlankPattern() {
        assertThrows(PrefixRegistrationRequestValidatorException.class, () ->
                validator.validate(fnc.apply(""))
        );
    }

    @Test
    void validateNotAPattern() {
        assertThrows(PrefixRegistrationRequestValidatorException.class, () ->
                validator.validate(fnc.apply("[]"))
        );
    }

    @Test
    void validateValidPattern() {
        assertTrue(validator.validate(fnc.apply("\\d+")));
    }
}